LabelledSeriesValueEnumerable
