+++ src/DSE.Open.Numerics/IReadOnlySeries.T.cs
@@ -15,6 +15,19 @@ public interface IReadOnlySeries<T>
     where T : IEquatable<T>
 {
     /// <summary>
+    /// Provides labels for data values in the series. There is no guarantee that every data value is
+    /// labelled, nor that every label is associated with a data value.
+    /// </summary>
+    IReadOnlyDataLabelCollection<T> Labels { get; }
+
+    /// <summary>
+    /// Gets a read-only view of the labelled values in the series. Values for which there is no
+    /// corresponding label in <see cref="Labels"/> are labelled by calling <see cref="object.ToString()"/>
+    /// on the value.
+    /// </summary>
+    IEnumerable<string> LabelledValues { get; }
+
+    /// <summary>
     /// Gets a span over the contents of the vector.
     /// </summary>
     ReadOnlySpan<T> AsReadOnlySpan();
